+++ tests/important-boolean.test.js
@@ -8,7 +8,7 @@ crosscheck(() => {
   test('important boolean', () => {
     let config = {
       important: true,
-      darkMode: 'class',
+      darkMode: 'selector',
       content: [
         {
           raw: html`
@@ -148,10 +148,10 @@ crosscheck(() => {
             text-align: right !important;
           }
         }
-        :is(:where([dir='rtl']) .rtl\:active\:text-center:active) {
+        .rtl\:active\:text-center:active:where([dir='rtl'], [dir='rtl'] *) {
           text-align: center !important;
         }
-        :is(:where(.dark) .dark\:focus\:text-left:focus) {
+        .dark\:focus\:text-left:focus:where(.dark, .dark *) {
           text-align: left !important;
         }
       `)