+++ tests/important-modifier-prefix.test.js
@@ -5,7 +5,7 @@ crosscheck(() => {
     let config = {
       important: false,
       prefix: 'tw-',
-      darkMode: 'class',
+      darkMode: 'selector',
       content: [
         {
           raw: html`<!-- The string "!*" can cause problems if we don't handle it, let's include it -->