+++ tests/kitchen-sink.test.js
@@ -3,7 +3,7 @@ import { crosscheck, run, html, css, defaults } from './util/run'
 crosscheck(({ stable, oxide }) => {
   test('it works', () => {
     let config = {
-      darkMode: 'class',
+      darkMode: 'selector',
       content: [
         {
           raw: html`
@@ -304,8 +304,10 @@ crosscheck(({ stable, oxide }) => {
           margin-right: auto;
         }
         .drop-empty-rules:hover,
-        .group:hover .apply-group,
-        :is(:where(.dark) .apply-dark-mode) {
+        .group:hover .apply-group {
+          font-weight: 700;
+        }
+        .apply-dark-mode:where(.dark, .dark *) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -340,7 +342,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        :is(:where(.dark) .group:hover .apply-dark-group-example-a) {
+        .group:hover .apply-dark-group-example-a:where(.dark, .dark *) {
           --tw-bg-opacity: 1;
           background-color: rgb(34 197 94 / var(--tw-bg-opacity));
         }
@@ -802,12 +804,12 @@ crosscheck(({ stable, oxide }) => {
             text-align: left;
           }
         }
-        :is(:where(.dark) .dark\:custom-util) {
+        .dark\:custom-util:where(.dark, .dark *) {
           background: #abcdef;
         }
         @media (min-width: 768px) {
           @media (prefers-reduced-motion: no-preference) {
-            :is(:where(.dark) .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
+            .md\:dark\:motion-safe\:foo\:active\:custom-util:active:where(.dark, .dark *) {
               background: #abcdef !important;
             }
           }
@@ -877,8 +879,10 @@ crosscheck(({ stable, oxide }) => {
           margin-right: auto;
         }
         .drop-empty-rules:hover,
-        .group:hover .apply-group,
-        :is(:where(.dark) .apply-dark-mode) {
+        .group:hover .apply-group {
+          font-weight: 700;
+        }
+        .apply-dark-mode:where(.dark, .dark *) {
           font-weight: 700;
         }
         .apply-with-existing:hover {
@@ -912,7 +916,7 @@ crosscheck(({ stable, oxide }) => {
         .apply-order-b {
           margin: 1.5rem 1.25rem 1.25rem;
         }
-        :is(:where(.dark) .group:hover .apply-dark-group-example-a) {
+        .group:hover .apply-dark-group-example-a:where(.dark, .dark *) {
           background-color: #22c55e;
         }
         @media (min-width: 640px) {
@@ -1364,12 +1368,12 @@ crosscheck(({ stable, oxide }) => {
             text-align: left;
           }
         }
-        :is(:where(.dark) .dark\:custom-util) {
+        .dark\:custom-util:where(.dark, .dark *) {
           background: #abcdef;
         }
         @media (min-width: 768px) {
           @media (prefers-reduced-motion: no-preference) {
-          :is(:where(.dark) .md\:dark\:motion-safe\:foo\:active\:custom-util:active) {
+          .md\:dark\:motion-safe\:foo\:active\:custom-util:active:where(.dark, .dark *) {
             background: #abcdef !important;
           }
         }